+++ src/room/index.ts
@@ -1,4 +1,4 @@
-import Model from './model';
+import Model from '../model';
 
 export default class Room extends Model<Room> {
   id: string;
