+++ src/supply/index.ts
@@ -1,4 +1,4 @@
-import Model from './model';
+import Model from '../model';
 
 export default class Supply extends Model<Supply> {
   id: string;
