+++ tsconfig.json
@@ -7,7 +7,7 @@
     "noImplicitAny": true,
     "noImplicitReturns": true,
     "suppressImplicitAnyIndexErrors": false,
-    "target": "ES6"
+    "target": "ES5"
   },
   "exclude": [
     "node_modules"
